requirements.txt
Django==1.10.3
django-cors-middleware==1.3.1
django-filter==1.0.1
djangorestframework==3.5.3
drf-jsonapi==0.1.6
inflection==0.3.1
